feat: redesign gameplay UI
User prompt: For the gameplay, I’d like to update the design a bit. I’d like to show only the current word being guessed as well as use input boxes that correspond to the number of missing letters. As players type each letter, I’d like to move to the next box automatically. They should still be able to tab through the boxes as they normally would. We will also have a heads up for how many words they have guessed, how many words their opponent has guessed, and perhaps hidden tabs that can be popped open showing their wrong guesses and the opponents wrong guesses. If the guess is right, transition to a check mark in the centre of the screen before transitioning to the next word (ie with the boxes again)