Update deployment workflow docs
The GitHub workflow for deploying the server via ssh appears to overwrite the entire target directory, this causes important files such as authorised ssh keys that are stored there to be lost upon each deploy. Please confirm that this is the case and if so propose a fix.

You can do all this while still using the web-deploy action. It has an `rsync-options` parameter where you can pass specific options to the underlying rsync command. I think we should name the target folder `server` otherwise the approach makes sense to me

I think we can make the deployment doc more generic, ie addressing what the current flow is, rather than what it was. Also, I think now that we’ve changed the destination, we can remove the rsync options. The defaults should be good enough.