feat: surface completed and upcoming words in game board
User prompt: Let’s make some minor UI changes. Let’s show the previously completed words to the user. Let’s also show the words that are next (with the blocks obviously). The blocks also seem to be the different sizes for the editable and non-editable. For when a player is waiting for the opponent to complete their turn, let’s make it more obvious. Perhaps some styling to show that the inputs are disabled.