Add 20s turn timer
Let’s add a timer feature to the game. I’d like to have 20 seconds per turn for each player. At the end of the 20 seconds, if the player hasn’t made a guess, their turn will be considered failed. Upon a successful guess, their timer will reset to the 20 sends once more. The timer should be visible in a fixed position on the top right of the screen at all times.